refactored container functions to handle single org verifications
